Add files via upload
[0:30]

Welcome back, viewers! Today, we're diving into our second step: Configuration Management. We're using Burp Suite to crawl the website and uncover numerous URLs. In the intercept section, we found a significant GET request at /index.php.

[1:15]

Sending this request to the Intruder, we marked index.php and set up payloads. We have one directory already set up in Burp Suite. Creating a resource pool here, we initiated our attack. As you can see, we received a 404 status, indicating that something interesting might be happening. Filtering out this intruder, we discovered the cgi-bin endpoint, but it's blocked with a 403 error. We also located the admin endpoint.

[1:31]

Now, we're exploring the admin endpoint. Upon entering, we found a create.sol file. Try downloading this file and checking for SQL injection. Don't worry; I'll guide you through it. It's essential for your practice.

[4:05]

Moving on, we tested cgi-bin accessibility. Unfortunately, it's forbidden, and port 8000 is also inaccessible.

[4:45]

Burp Suite identified a directory listing at [6:04], so we sent the index.php request to Burp Intruder. We set up payloads using a simple list and initiated the attack. However, most of the website responses were '404 not found.'

[7:13]

Next, we assessed the website's security headers. For this, we used the SecurityHeader website. Running a scan on our site, we found a few headers missing: Content-Security-Policy, X-Frame-Options, X-Content-Type-Options, Referrer-Policy. This is a crucial area to work on for enhancing our security posture.

[9:25]

Further testing involved checking policies such as Flash, Silverlight, and robots. By altering .php to .swf and inspecting the page source, we identified a JavaScript file. Copy-pasting it into Notepad, we analyzed its contents.

[10:05]

Lastly, we explored various lists. There are many options, but we chose a shorter one for efficiency. These lists are vital for monetization and further exploitation.